prevent metering overflow
use forward 64bit to RRD, use double internally

refs #2476

+++ src/model/modulator.cpp
@@ -174,7 +174,7 @@ namespace dss {
     return m_PowerConsumptionTimeStamp;
   }
 
-  unsigned long DSMeter::getCachedEnergyMeterValue() {
+  double DSMeter::getCachedEnergyMeterValue() {
     return m_EnergyMeterValue;
   } // getEnergyMeterValue
